Code Cleanup done

+++ lib/model/api_response_model.dart
@@ -10,7 +10,7 @@ class APIResponseModel {
   int casesPerOneMillion;
   int deathsPerOneMillion;
   int tests;
-  double testsPerOneMillion;
+  int testsPerOneMillion;
   int affectedCountries;
 
   APIResponseModel(
